[Android] Fix race condition dispatching last intent
In some edge cases, the delay between checking the lastPushIntent being
null and calling the handler would be enough time for the lastPushIntent
to be modified (or set to null). The best option here seems to be to
store a local copy of the intent at the time we check it, and then call
the handler with that local copy rather than a potentially modified
intent.